Support enlish dish names
As english translations for dishes are not guaranteed, there is
a fallback to the german version of the dish's name in place.
The alt-text of the dish's image is also set to the correct version.

There may be further considerations needed on how to visualize
the missing native language value.