fix: Hide occurrences containing the same dish id for a given day
The bug resulting in multiple occurrences for a single day was
already fixed on the scraper side. It would be best to retroactively
clean up the database and remove duplicate entries, but as this is
a risky process we should avoid it for now.

After we prune the entires, this workaround can be removed.